+++ packages/vendure-plugin-invoices/src/services/invoice.service.ts
@@ -22,6 +22,7 @@ import {
   RequestContext,
   TransactionalConnection,
   UserInputError,
+  EntityRelationPaths,
 } from '@vendure/core';
 import {
   Invoice,
@@ -76,6 +77,11 @@ export class InvoiceService implements OnModuleInit, OnApplicationBootstrap {
       }>
     | undefined;
   retries = 10;
+  orderRelations: EntityRelationPaths<Order>[] = [
+    'lines.productVariant.product',
+    'shippingLines.shippingMethod',
+    'payments',
+  ];
 
   constructor(
     private eventBus: EventBus,
@@ -304,7 +310,7 @@ export class InvoiceService implements OnModuleInit, OnApplicationBootstrap {
   ): Promise<InvoiceEntity | undefined> {
     const ctx = await this.createCtx(channelToken);
     let [order, previousInvoiceForOrder, config] = await Promise.all([
-      this.orderService.findOneByCode(ctx, orderCode),
+      this.orderService.findOneByCode(ctx, orderCode, this.orderRelations),
       this.getMostRecentInvoiceForOrder(ctx, orderCode),
       this.getConfig(ctx),
     ]);
@@ -452,7 +458,7 @@ export class InvoiceService implements OnModuleInit, OnApplicationBootstrap {
       ctx,
       new Injector(this.moduleRef),
       order,
-      latestInvoiceNumber,
+      latestInvoiceNumber ?? this.config.startInvoiceNumber,
       shouldGenerateCreditInvoice
     );
     const tmpFilePath = await createTempFile('.pdf');
@@ -497,7 +503,11 @@ export class InvoiceService implements OnModuleInit, OnApplicationBootstrap {
   ): Promise<ReadStream> {
     let order: Order | undefined;
     if (orderCode) {
-      order = await this.orderService.findOneByCode(ctx, orderCode);
+      order = await this.orderService.findOneByCode(
+        ctx,
+        orderCode,
+        this.orderRelations
+      );
     } else {
       const orderId = (
         await this.orderService.findAll(
@@ -509,7 +519,11 @@ export class InvoiceService implements OnModuleInit, OnApplicationBootstrap {
           []
         )
       )?.items[0].id;
-      order = await this.orderService.findOne(ctx, orderId);
+      order = await this.orderService.findOne(
+        ctx,
+        orderId,
+        this.orderRelations
+      );
     }
     if (!order) {
       throw new UserInputError(`No order found with code ${orderCode}`);